Updated DB handling
In order for the client to receive the lastest message in each conversation, I had to find a workaround from the previous aggregate() function that was used. The previous implementation was only retrieving the last message from the sender, and not the last message from the discussion.

The solution was to create a separate collection in mongodb that will only save the last message, and keeps it up to date each time someone interact in the conversation.