Created a passphrase login
All the data needed for a user account to work with our service is now stored in an encrypted state in the localStorage. I added a passphrase selector screen, and a passphrase login screen to unlock the storage and access your encryption key, authentication token, etc...

The code was also cleared a little by integrating the UI.js module.