Create STMServer.js
This module streams the blockchain, and search for incoming transaction/vote to the steem-messenger account/last steem-messenger post.

This script will also save all the users that send a transaction/vote into the database (`db.setSubscription`). That way, we can now check on the server side if a given user has a paid plan or not.