Complete redesign of the client logic
In order to implement the new `safeSocket` module, we needed to rethink the way the client/server communications works. This will also be the starting point to implementing the api method.

We also reduced the many eventListeners functions to only one, in a seamless point&click app style. The code detects where a user clicked, and reacts accordingly. The click events's data are not saved and aren't sent to any party