Complete Steem Messenger server redesign
The logic of the code was heavily updated. 

Previously, there was a socket.emit method for each call the client/server would make. While this is not a major problem in itself as it worked, it posed some upgradability problems, and some security issues, a not all socket methods were secured by an authentication token.

Therefore, we added the Lara's `encodeSafeSocket` function, which prepare the data in a particular format (json object), encrypt it, and it then returns to a socket.emit "safeSocket" method. This way, attackers will have a hard time guessing how to differenciate a request from another, and thus reducing potentials network analysis.

This will also give us the foundation to add our antispam algorithm