Update serverlib.js
Added "add to blacklist" event case. The function `addToBlacklist` will save the blacklisted user in the dabatase. The blacklisted user will not be capable to flood/spam the user that blacklisted him, because the server checks in the `handleInput` function if the user has been blacklisted by his recipient. If this is the case, no communication is sent to the recipient, avoiding any spam attack for our users