Finalized the GCM implementation
We implemented AES-GCM for user-user communications, but the old steem.memo.encode/decode was still used in the safeSocket method, which was a huge bottleneck. 

We also revised how the session keys were managed on the client side. Now, session keys are generated at each login, and not each interaction as before, which was also a bottleneck (too many api calls, + too many DH calls)